requirements.txt
apache-bookkeeper-client==4.13.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
coloredlogs==15.0
fastavro==1.3.5
greenlet==1.0.0
grpcio==1.27.2
humanfriendly==9.1
idna==2.10
importlib-metadata==3.10.0
jsonschema==3.2.0
numpy==1.20.2
pandas==1.1.5
prometheus-client==0.10.1
protobuf==3.15.8
pulsar-client==2.6.3
pymmh3==0.0.5
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
ratelimit==2.2.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.7
typing-extensions==3.7.4.3
urllib3==1.26.4
zipp==3.4.1
